discord_bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.30.0
solana>=0.30.2
requests>=2.31.0
